chore: Extract app and coordinator event handlers to respective crates
Move relevant configs as well.

They were only there because ln_dlc_node_tests dependended on the configurations.
