+++ coordinator/Cargo.toml
@@ -4,6 +4,7 @@ version = "1.3.0"
 edition = "2021"
 
 [dependencies]
+async-trait = "0.1"
 atty = "0.2.14"
 bitcoin = "0.29"
 console-subscriber = "0.1.6"
