+++ crates/ln-dlc-node-tests/Cargo.toml
@@ -10,13 +10,15 @@ edition = "2021"
 [dev-dependencies]
 anyhow = "1"
 bitcoin = "0.29"
-ln-dlc-node = { path = "../ln-dlc-node" }
+coordinator = { path = "../../coordinator" }
 dlc-manager = { version = "0.4.0", features = ["use-serde"] }
 dlc-messages = { version = "0.4.0" }
 dlc-trie = { version = "0.4.0" }
 futures = "0.3"
 hex = "0.4"
 lightning = { version = "0.0.114", features = ["max_level_trace"] }
+ln-dlc-node = { path = "../ln-dlc-node" }
+native = { path = "../../mobile/native" }
 rand = "0.8.5"
 reqwest = { version = "0.11", default-features = false, features = ["json"] }
 rust_decimal = "1"
