+++ crates/ln-dlc-node-tests/src/tests/mod.rs
@@ -1,23 +1,3 @@
-use bitcoin::hashes::sha256;
-use dlc_messages::Message;
-use ln_dlc_node::HTLCStatus;
-use ln_dlc_node::config::app_config;
-use ln_dlc_node::config::coordinator_config;
-use ln_dlc_node::config::LIQUIDITY_MULTIPLIER;
-use ln_dlc_node::node::InMemoryStore;
-use ln_dlc_node::node::LnDlcNodeSettings;
-use ln_dlc_node::node::Node;
-use ln_dlc_node::node::NodeInfo;
-use ln_dlc_node::node::OracleInfo;
-use ln_dlc_node::node::RunningNode;
-use ln_dlc_node::node::sub_channel_message_name;
-use ln_dlc_node::scorer;
-use ln_dlc_node::seed::Bip39Seed;
-use ln_dlc_node::util;
-use ln_dlc_node::AppEventHandler;
-use ln_dlc_node::CoordinatorEventHandler;
-use ln_dlc_node::EventHandlerTrait;
-use ln_dlc_node::EventSender;
 use anyhow::Result;
 use bitcoin::hashes::sha256;
 use bitcoin::secp256k1::PublicKey;
@@ -45,6 +25,22 @@ use futures::Future;
 use lightning::ln::channelmanager::ChannelDetails;
 use lightning::util::config::UserConfig;
 use lightning::util::events::Event;
+use ln_dlc_node::config::LIQUIDITY_MULTIPLIER;
+use ln_dlc_node::node::sub_channel_message_name;
+use ln_dlc_node::node::InMemoryStore;
+use ln_dlc_node::node::LnDlcNodeSettings;
+use ln_dlc_node::node::Node;
+use ln_dlc_node::node::NodeInfo;
+use ln_dlc_node::node::OracleInfo;
+use ln_dlc_node::node::RunningNode;
+use ln_dlc_node::scorer;
+use ln_dlc_node::seed::Bip39Seed;
+use ln_dlc_node::util;
+use ln_dlc_node::EventHandlerTrait;
+use ln_dlc_node::EventSender;
+use ln_dlc_node::HTLCStatus;
+use native::config::app_config;
+use native::ln_dlc::app_event_handler::AppEventHandler;
 use rand::distributions::Alphanumeric;
 use rand::thread_rng;
 use rand::Rng;
