+++ crates/ln-dlc-node/Cargo.toml
@@ -43,4 +43,3 @@ tracing = "0.1.37"
 tracing-log = "0.1.3"
 ureq = "2.5.0"
 uuid = { version = "1.3.0", features = ["v4", "serde"] }
-
