+++ crates/ln-dlc-node/src/lib.rs
@@ -46,9 +46,7 @@ pub use config::LIQUIDITY_MULTIPLIER;
 pub use ldk_node_wallet::WalletSettings;
 pub use lightning;
 pub use lightning_invoice;
-pub use ln::AppEventHandler;
 pub use ln::ChannelDetails;
-pub use ln::CoordinatorEventHandler;
 pub use ln::DlcChannelDetails;
 pub use ln::EventHandlerTrait;
 pub use ln::EventSender;
