+++ crates/ln-dlc-node/src/ln/mod.rs
@@ -1,6 +1,4 @@
-mod app_event_handler;
 mod channel_details;
-mod coordinator_event_handler;
 mod dlc_channel_details;
 mod event_handler;
 mod logger;
@@ -13,11 +11,11 @@ mod manage_spendable_outputs;
 /// don't require custom behaviour.
 pub mod common_handlers;
 
-pub use app_event_handler::AppEventHandler;
 pub use channel_details::ChannelDetails;
-pub use coordinator_event_handler::CoordinatorEventHandler;
 pub use dlc_channel_details::DlcChannelDetails;
 pub use event_handler::EventHandlerTrait;
 pub use event_handler::EventSender;
+pub use event_handler::InterceptionDetails;
+pub use event_handler::PendingInterceptedHtlcs;
 pub(crate) use logger::TracingLogger;
 pub(crate) use manage_spendable_outputs::manage_spendable_outputs;
